--- new-docs/puppeteer.puppeteer._projectroot.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [puppeteer](./puppeteer.md) &gt; [Puppeteer](./puppeteer.puppeteer.md) &gt; [\_projectRoot](./puppeteer.puppeteer._projectroot.md)
-
-## Puppeteer.\_projectRoot property
-
-<b>Signature:</b>
-
-```typescript
-_projectRoot: string;
-```
