+++ scripts/test-install.sh
@@ -22,3 +22,18 @@ PUPPETEER_PRODUCT=firefox npm install --loglevel silent "${tarball}"
 node --eval="require('puppeteer')"
 rm "${tarball}"
 ls $TMPDIR/node_modules/puppeteer/.local-firefox/linux-79.0a1/firefox/firefox
+
+# Again for puppeteer-core
+node ./utils/prepare_puppeteer_core.js
+npm pack
+tarball="$(realpath puppeteer-core-*.tgz)"
+TMPDIR="$(mktemp -d)"
+cd $TMPDIR
+# Check we can install from the tarball.
+# This emulates installing from npm and ensures that:
+# 1. we publish the right files in the `files` list from package.json
+# 2. The install script works and correctly exits without errors
+# 3. Requiring Puppeteer Core from Node works.
+npm install --loglevel silent "${tarball}"
+node --eval="require('puppeteer-core')"
+
