+++ src/common/Puppeteer.ts
@@ -23,8 +23,6 @@ import { ProductLauncher } from '../node/Launcher';
 import { BrowserFetcher, BrowserFetcherOptions } from '../node/BrowserFetcher';
 import { puppeteerErrors, PuppeteerErrors } from './Errors';
 import { ConnectionTransport } from './ConnectionTransport';
-import readPkgUp from 'read-pkg-up';
-
 import { devicesMap } from './DeviceDescriptors';
 import { DevicesMap } from './DeviceDescriptors';
 import { Browser } from './Browser';
@@ -35,6 +33,7 @@ import {
   clearQueryHandlers,
   QueryHandler,
 } from './QueryHandler';
+import { PUPPETEER_REVISIONS } from '../revisions';
 
 /**
  * The main Puppeteer class
@@ -58,7 +57,7 @@ import {
  * @public
  */
 export class Puppeteer {
-  _projectRoot: string;
+  private _projectRoot: string;
   _preferredRevision: string;
   _isPuppeteerCore: boolean;
   _changedProduct = false;
@@ -173,14 +172,13 @@ export class Puppeteer {
       this._lazyLauncher.product !== this._productName ||
       this._changedProduct
     ) {
-      const { packageJson } = readPkgUp.sync({ cwd: __dirname });
       switch (this._productName) {
         case 'firefox':
-          this._preferredRevision = packageJson.puppeteer.firefox_revision;
+          this._preferredRevision = PUPPETEER_REVISIONS.firefox;
           break;
         case 'chrome':
         default:
-          this._preferredRevision = packageJson.puppeteer.chromium_revision;
+          this._preferredRevision = PUPPETEER_REVISIONS.chromium;
       }
       this._changedProduct = false;
       this._lazyLauncher = Launcher(
