+++ src/index-core.ts
@@ -16,5 +16,5 @@
 
 import { initializePuppeteer } from './initialize';
 
-const puppeteer = initializePuppeteer('puppeteer');
+const puppeteer = initializePuppeteer('puppeteer-core');
 export default puppeteer;
