+++ src/revisions.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2017 Google Inc. All rights reserved.
+ * Copyright 2020 Google Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,12 @@
  * limitations under the License.
  */
 
-import { initializePuppeteer } from './initialize';
+type Revisions = Readonly<{
+  readonly chromium: string;
+  readonly firefox: string;
+}>;
 
-const puppeteer = initializePuppeteer('puppeteer');
-export default puppeteer;
+export const PUPPETEER_REVISIONS: Revisions = {
+  chromium: '756035',
+  firefox: 'latest',
+};
