+++ utils/prepare_puppeteer_core.js
@@ -23,4 +23,5 @@ const json = require(packagePath);
 
 json.name = 'puppeteer-core';
 delete json.scripts.install;
+json.main = './cjs-entry-core.js';
 fs.writeFileSync(packagePath, JSON.stringify(json, null, '  '));
